networks can be removed, old code cleanup
